Make JumpNotZero generic.
