+++ crates/sierra/src/extensions/core.rs
@@ -1,3 +1,4 @@
+use super::conditional_jump::JumpNotZeroLibFunc;
 use super::modules::felt::{FeltLibFunc, FeltType};
 use super::modules::function_call::FunctionCallLibFunc;
 use super::modules::gas::{GasBuiltinType, GasLibFunc};
@@ -22,6 +23,7 @@ define_libfunc_hierarchy! {
         FunctionCall(FunctionCallLibFunc),
         Gas(GasLibFunc),
         Integer(IntegerLibFunc),
+        JumpNotZero(JumpNotZeroLibFunc),
         Mem(MemLibFunc),
         UnwrapNonZero(UnwrapNonZeroLibFunc),
         UnconditionalJump(UnconditionalJumpLibFunc),
