+++ crates/sierra/src/simulation/core.rs
@@ -8,9 +8,7 @@ use crate::extensions::core::CoreConcreteLibFunc::{
 use crate::extensions::function_call::FunctionCallConcreteLibFunc;
 use crate::extensions::gas::GasConcreteLibFunc::{GetGas, RefundGas};
 use crate::extensions::gas::{GetGasConcreteLibFunc, RefundGasConcreteLibFunc};
-use crate::extensions::integer::IntegerConcrete::{
-    Const, Duplicate, Ignore, JumpNotZero, Operation,
-};
+use crate::extensions::integer::IntegerConcrete::{Const, Duplicate, Ignore, Operation};
 use crate::extensions::integer::{
     BinaryOperationConcreteLibFunc, ConstConcreteLibFunc, OperationConcreteLibFunc,
     OperationWithConstConcreteLibFunc, Operator,
@@ -19,6 +17,7 @@ use crate::extensions::mem::MemConcreteLibFunc::{
     AlignTemps, AllocLocals, Rename, StoreLocal, StoreTemp,
 };
 use crate::ids::FunctionId;
+use crate::simulation::CoreConcreteLibFunc::JumpNotZero;
 
 /// Simulates the run of a single libfunc. Returns the memory reperesentations of the outputs, and
 /// the chosen branch given the inputs. A function that provides the simulation of running a user
@@ -98,7 +97,7 @@ pub fn simulate<
             unpack_inputs::<1>(inputs)?;
             Ok((vec![], 0))
         }
-        Integer(JumpNotZero(_)) => {
+        JumpNotZero(_) => {
             let [MemCell { value }] = unpack_inputs::<1>(inputs)?;
             if value != 0 {
                 // Non-zero - jumping to the success branch and providing a NonZero wrap to the
